+++ packages/cli/src/infrastructure/console/progress.rs
