+++ packages/cli/src/infrastructure/console/terminal.rs
