+++ packages/core/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod application;
 pub mod config;
 pub mod domain;
-pub mod infra;
+pub mod infrastructure;
 pub mod utils;
 pub use utils::*;
